Conforms service to dependency injection system
