+++ Modules/Sources/DashRepository/ForzaService.swift
@@ -12,7 +12,7 @@ public struct ForzaService{
 }
 
 extension ForzaService: TestDependencyKey {
-    public static let testValue = Self(getForzaInfo: unimplemented("getForzaInfo"))
+    public static let testValue: ForzaService = .mock
 }
 
 extension DependencyValues {
@@ -21,3 +21,21 @@ extension DependencyValues {
         set { self[ForzaService.self] = newValue }
     }
 }
+
+#if DEBUG
+public extension ForzaService {
+    static var counter = -1
+    static let mock: Self = .init {
+        Timer.publish(every: 0.0001, on: .main, in: .default).autoconnect()
+            .map { _ in
+                if counter == 1000 {
+                    counter = -1
+                }
+                counter += 1
+                return ForzaModel.fixture(gameIsRunning: true, maxRPM: 1000, currentEngineRPM: Float(counter), speed: counter)
+            }
+            .setFailureType(to: Error.self)
+            .eraseToAnyPublisher()
+    }
+}
+#endif
