+++ Modules/Sources/DashRepositoryLive/ForzaService+Live.swift
@@ -1,4 +1,3 @@
-import ComposableArchitecture
 import Combine
 import Dependencies
 import Foundation
@@ -10,6 +9,7 @@ public extension ForzaService {
         UDPConnectionProvider
             .shared
             .getTelemetryData()
+            .receive(on: DispatchQueue.main)
             .map { response in
                 ForzaModel(
                     gameIsRunning: response.gameIsRunning,
@@ -29,7 +29,8 @@ public extension ForzaService {
                     carClass: response.carClass,
                     carPerformanceIndex: response.carPerformanceIndex,
                     driveTrainType: response.driveTrainType,
-                    numOfCylinders: response.numOfCylinders
+                    numOfCylinders: response.numOfCylinders,
+                    distanceTraveled: response.distanceTraveled
                 )
             }
             .mapError { error in
@@ -42,22 +43,3 @@ public extension ForzaService {
 extension ForzaService: DependencyKey {
     public static var liveValue: ForzaService = .live
 }
-
-
-#if DEBUG
-public extension ForzaService {
-    static var counter = -1
-    static let mock: Self = .init {
-        Timer.publish(every: 0.1, on: .main, in: .default).autoconnect()
-            .map { _ in
-                if counter == 100 {
-                    counter = -1
-                }
-                counter += 1
-                return ForzaModel.fixture(gameIsRunning: true, speed: counter)
-            }
-            .setFailureType(to: Error.self)
-            .eraseToAnyPublisher()
-    }
-}
-#endif
